Move onResize logic to separate hook
